basic: add floating point arithmetic option
This sorta works, except parsing of FP constants is not implemented.

Arithmetics are about 5 or 6 times slower than with integers. A
performance drop was to be expected, but this is rather high, considering
that the general interpretation overhead should not be affected by this
change. Furthermore, enabling this option slows down drawing of tiled
backgrounds considerably, even though that doesn't even use any of the
interpreter code.

This is even more mysterious if one considers that the very significant
slowdown (16ms vs 10ms for rendering a background with sprites) happens
in code that largely spends its time waiting for the video controller...