psx: avoid polling non-existent controller (untested)
If polling the controller has failed, we will not attempt to do so again
unless there is a pause of at least one second between read attempts.

This saves a lot of wasted time checking for a controller that isn't
there, but still allows plugging in one without resetting the machine.