basic: seed RNG properly
Without this, the RNG is not completely predictable, but it's not really
random either.

At this point, we have synchronized the timer with the VS23 video
controller and have initialized the keyboard, which are AFAICT the only
things in the boot sequence that depend on anything outside the ESP8266.
This makes the cycle counter as good a seed as it gets.