ttconfig.h: prepare for memory layout strategy reversal
ATM, BASIC core functions are grouped together in flash. This is not
fool-proof for two reasons:
- libc functions (FP math in particular) are in flash, but not grouped
- various data arrays are in flash, but in different sections

Avoiding the first condition in particular requires intrusive modifications
to the BSP, which I am not prepared to do.

Instead, we move the entire BASIC core to RAM and move the I/O functions
(which are currently occupying most of IRAM) to flash.

Grouping RO data together with those functions in flash, using a slightly
modified linker script, avoids conflicts between both those and the BASIC
core functions.