nosdki2s: re-use BIOS print buffer as I2S buffer (super-experimental)
Saves 2.5K of RAM, but makes three big assumptions:

1. The buffer is not used. This seems to be valid, considering that the
Arduino core uses its own printf() and sprintf(), and that debug output
does not seem to affect sound quality (or crash the system).

2. There is no other purpose for the memory between print_mem_buf and
the stack, because the sound buffers are bit larger (2560 bytes) than the
print buffer (1820 bytes). This is somewhat difficult to verify.

3. We don't need excessive amounts of stack. If 2. is true, we have about
6k of stack by default, and about 5k when using the print buffer for I2S.
The only thing we know for sure at this time is that we need more than 2k.