vs23: dump redundant "transparent" property
An opaque sprite does not need a pattern in RAM, so s->pat == NULL is
the same as s->p.transparent. Dump the flag and reset s->pat if we find
that a sprite is opaque.

XXX: Still keeps the analyzed pattern in the cache until the next
garbage collection.