basic: putlist(): do not print spaces between keywords and '('
This is to avoid spaces between functions and their arguments; this used
to be done using the nsa table, but that won't work for keywords that
double as commands or inner keywords.

The catch is that stuff like "SPRITE 0 FRAME (x+y)*2" will be printed
as "SPRITE 0 FRAME(x+y)*2", which makes it look as if the FRAME() function
is used, when it is in fact the sprite's FRAME parameter.