basic: basic(): initialize SD card before VS23
The MicroSD card I use for development (SanDisk 512MB) has a habit of
talking on the SPI bus without being asked to while it is not initialized
yet. This can lead to all sorts of data corruption while setting up the
video chip, with results ranging from small artifacts to a completely
garbled display.

This change sets up the SD card before the video chip. The slight downside
is that if there is no SD card inserted, booting does not show anything
on screen for a few more seconds than usual.