vs23: setMode(): add a pause after mode switching for TV to sync
Doing stuff directly after a new mode has been set (especially drawing
a colored border) seems to confuse the KX-14CP1 (and potentially other
TVs/monitors as well). To avoid that, we wait a couple of frames before
returning from setMode() to give the display time to synchronize to the
new mode.