tscreenBase, tTVscreen: retool text window handling
Instead of dumping and reallocating the text memory every time a window
changes, retain it and keep track of the window dimensions separately
from the whole-screen dimensions.

This allows a window to retain its contents when switching to another
one without losing consistency of the text memory with what is actually
displayed on screen.

Makes windows usable for things other than error messages in bg/sprite
programs.