basic: CALL: fix yet another stack frame bug (with test)
We cannot modify the stack indices while evaluating arguments because
that would break references to local variables of the caller. That
means, however, that we also cannot push evaluated parameters on the
stack because they may be overwritten by functions called while evaluating
subsequent parameters.

We avoid this by storing parameters in temporary lists.

XXX: Probably not good for performance...