basic: SCREEN: set font after resetting screen
Unevenly sized fonts clip the text window by a few pixels. When the text
screen is reset these window dimensions are preserved, and when a font
that fits the screen resolution evenly is selected afterwards, it will
be missing the clipped pixels and the screen will have one character less
per line.

Fixes lost characters when using SCREEN with the same mode already active
and a non-default font selected.