basic: interpret numeric variable without = as unknown command
A string of characters that is not a keyword and not followed by an
= operator is most likely a misspelled command rather than a malformed
numeric variable assignment, so we treat it as such so as not to confuse
beginners, who are very likely to see this a lot.

Strings, lists, etc. have a sigil of some sort that clearly identifies
them as variables, so we retain the "= expected" error in those cases.