basic: re-implement slowdown key handling
Doing it in newline() breaks when text is output that does not contain
any newlines.

Doing it in screen_putch() fixes that, and we don't have to block for a
long time there, and thus have no need for various workarounds that abort
the delay if something happens during the pause.