basic: make BASIC context a proper class
Removes the horrible context macros and improves performance slightly.

This is less intrusive than it looks. Functional differences are limited
to:
- the tscreenBase class (which has to use the BASIC context explicitly now,
and also needs to deal with the case that there is none),
- the EXEC handler (which has to construct and use its subprogram context
differently),
- the pump_events() function, which is really a part of the OS and also
needs to use an explicit BASIC context now, and
- the main() function, which is now in charge of constructing the initial
BASIC context.

The remaining changes are mostly adjustments to accommodate C++ syntax
quirks, especially in the icode.py script.

Some things that are yet to be done:
- Move some leftover BASIC status into the class.
- Have a proper constructor and descructor.

There is no useful way in which to split this commit because all changes
depend on each other.