hosted: Arduino: hack to avoid having to re-implement stdio
chroot() to virtual Engine BASIC file system so all the unistd/stdio
functions work the same as on a real target system. This means that
the hosted build needs to be run as root (ouch), and that the sample
rate cannot be changed because SDL can't find its device files anymore.

Still beats a swift kick in the face.