h3gfx: directly render to framebuffer(s)
When the BG engine is off, we directly render to the visible
framebuffer, using dcache flush to avoid "snow".

With the background engine enabled, double buffering is used. When a
"single-buffer" operation (pixel graphics, text) is performed, the
buffer that it is drawn into will be copied to avoid inconsistencies
that produce flickering text etc.