sdfiles: loadImage(), loadBitmap(): keep alpha when loading off-screen
The most likely use for loading bitmaps to off-screen memory is to use
them as sources for blits or the BG/sprite engine. In those cases it is
desirable to keep the alpha channel (or create one using the specified
mask) and discard what was already at that position in memory.

When loading to visible screen memory, OTOH, it is more likely that one
would want to blend the loaded image into the existing pixels.