basic: putlist(): fix incorrect spacing after REM, with test [backport!]
The generic spacing code looks at the next token to determine if there
should be a space or not. That doesn't work with I_REM (and I_SQUOT,
which already has an exception) because it is followed by a string
literal. The length icode of that string literal is then misinterpreted
as a token, with random results.