tscreenBase: fix more random crashes when changing screen modes
tscreenBase::init() is called with the maximum possible dimenstions for
a text screen, but depending on the current font's dimensions, not all
that area might actually be available. tmt writes to the screen and must
therefore be told the actually available screen area, or it will write
beyond the framebuffer, with predictably catastrophic results.

Observed on H3 when using a non-standard default screen mode, but can
really happen anytime, anywhere.