sdl: port keyboard handling to SDL2
SDL2 keyboard handling is garbage. Unicode symbols are not provided
anymore, and text input events are useless due to extreme lag.

This effectively means that the system keyboard layout and input methods
are unusable, and we have to do the same USB scancode translation as in
the bare metal H3 keyboard driver. Good job!