sdlgfx: move everything related to rendering to gfx thread
With SDL2, everything, literally _everything_ related to rendering has
to be done in the same thread, or it will randomly crap itself. Even the
initialization. And not just any thread, no, it has to be the main thread.
On some backends. Because otherwise it would be consistent, and we can't
have that, can we?