h3gfx: startCapture(): allocate zeroed capture buffers
The encoder requires dimensions to be multiples of 16, so 1080p has to
be extended by 8 lines. The RT-WB, however, does not write to those 8
lines, causing junk data (probably chroma) from the unused lines to
bleed into the real picture. Fixed by clearing the buffers on
allocation.

Fixes flickering line at the bottom of recordings of bright screens.