h3joy_sdl_client: handle axis 0/1 motion as directional input
Some controllers (notably the xbox360 controller, but also some generic
HID controllers) need to be switched to a different mode for the d-pad
to emit d-pad button events instead of analog axis motion events.

That is not at all obvious and may be cause for user confusion. To avoid it
we translate events on the analog axes 0 and 1 to d-pad events.

Fixes non-working d-pads on some mixed analog/digital controllers.