sdl: create/destroy windows in main thread
SDL2 is generally thread-unsafe, but it depends on the platform how that
is expressed.

It seems that while on Linux it's fine to create and destroy windows in
any thread, on OSX and Win32 it needs to be done in the main thread,
which is what this commit does.

Fixes hangs on startup on OSX and Win32.