sdlgfx: updateBgScale(): do not delay
This isn't necessary because SDL_RenderPresent() does it anyway.

This also partially works around an issue with either SDL or (more
likely) one or more Linux DRM drivers (observed with sun4i).

The DRM driver tends to skip a frame unless the flip is requested quite
early (max. 6 ms after vblank or so), effectively dropping the frame
rate to 30 fps. Since everything in EB is timed by the vblank, this
leads to a slowdown to 50%.

Inexplicably, this is only a problem when using SDL in double buffering
mode, not in (default) triple buffering mode.